:recycle: Extract Scraper.run again
