+++ src/execute.ts
@@ -15,37 +15,21 @@ export const execute = async (linesText: string): Promise<void> => {
       if (!line.trim()) continue;
 
       const mod = ModVersion.parse(line);
-      const b = new Builder(mod);
-      const patterns = b.patterns;
-
-      let success = false;
-      const errors: any[] = []; // eslint-disable-line @typescript-eslint/no-explicit-any
-      for (const ptn of patterns) {
-        try {
-          const license = await scraper.scrape(ptn.url, ptn.selector);
-          process.stdout.write(
-            JSON.stringify({
-              path: mod.path,
-              version: mod.version,
-              license,
-              url: ptn.url,
-            })
-          );
-          process.stdout.write('\n');
-          success = true;
-          break;
-        } catch (
-          err: any // eslint-disable-line @typescript-eslint/no-explicit-any
-        ) {
-          errors.push(err);
-        }
-      }
-      if (!success && errors.length > 0) {
-        process.stderr.write(
-          `failed to get license at ${patterns
-            .map(i => i.url)
-            .join(',')} because of ${errors}\n`
-        );
+      const patterns = new Builder(mod).patterns;
+      try {
+        const r = await scraper.run(patterns);
+        const license = {
+          path: mod.path,
+          version: mod.version,
+          license: r.license,
+          url: r.url,
+        };
+        process.stdout.write(JSON.stringify(license));
+        process.stdout.write('\n');
+      } catch (
+        err: any // eslint-disable-line @typescript-eslint/no-explicit-any
+      ) {
+        process.stderr.write(err.message);
       }
     }
   } finally {
